requirements.txt
Django==4.2.11
djangorestframework-simplejwt==5.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.8
whitenoise==6.6.0
openpyxl==3.1.2
pandas==2.2.2